+++ modules/apps/src/middleware.ts
@@ -6,11 +6,6 @@ import {
   UninstallMiddlewareContext,
   NetworkContext,
   ProposeMiddlewareContext,
-  InstallMiddlewareContext,
-  TakeActionMiddlewareContext,
-  CONVENTION_FOR_ETH_ASSET_ID,
-  ContractAddresses,
-  Address,
 } from "@connext/types";
 import { uninstallDepositMiddleware, proposeDepositMiddleware } from "./DepositApp";
 import { proposeLinkedTransferMiddleware } from "./SimpleLinkedTransferApp";
@@ -18,8 +13,8 @@ import { proposeHashLockTransferMiddleware } from "./HashLockTransferApp";
 import { proposeSignedTransferMiddleware } from "./SimpleSignedTransferApp";
 import { proposeWithdrawMiddleware } from "./WithdrawApp";
 import { proposeSwapMiddleware } from "./SimpleTwoPartySwapApp";
-import { commonAppProposalValidation, AppRegistry } from ".";
 import { stringify } from "@connext/utils";
+import { sharedProposalMiddleware } from "./shared/middleware";
 
 // add any validation middlewares
 export const generateValidationMiddleware = async (
@@ -34,23 +29,17 @@ export const generateValidationMiddleware = async (
     middlewareContext: MiddlewareContext,
   ) => {
     switch (protocol) {
-      case ProtocolNames.setup: {
-        break;
-      }
       case ProtocolNames.propose: {
         await proposeMiddleware(network, middlewareContext as ProposeMiddlewareContext);
         break;
       }
-      case ProtocolNames.install: {
-        await installMiddleware(network, middlewareContext as InstallMiddlewareContext);
-        break;
-      }
-      case ProtocolNames.takeAction: {
-        await takeActionMiddleware(network, middlewareContext as TakeActionMiddlewareContext);
-        break;
-      }
       case ProtocolNames.uninstall: {
         await uninstallMiddleware(network, middlewareContext as UninstallMiddlewareContext);
+        break;
+      }
+      case ProtocolNames.setup:
+      case ProtocolNames.install:
+      case ProtocolNames.takeAction: {
         break;
       }
 
@@ -79,33 +68,13 @@ const uninstallMiddleware = async (
   }
 };
 
-const getNameFromAddress = (contractAddress: ContractAddresses, appDefinition: Address) => {
-  const [name] =
-    Object.entries(contractAddress).find(([name, addr]) => addr === appDefinition) || [];
-  return name;
-};
-
 const proposeMiddleware = async (
   network: NetworkContext,
   middlewareContext: ProposeMiddlewareContext,
 ) => {
   const { contractAddresses } = network;
-  const { proposal, params } = middlewareContext;
-  const name = getNameFromAddress(contractAddresses, proposal.appDefinition);
-  // get registry information
-  const registryAppInfo = AppRegistry.find((app) => app.name === name);
-  if (!registryAppInfo) {
-    throw new Error(
-      `Refusing proposal of unsupported application. Cxt: ${stringify(middlewareContext)}`,
-    );
-  }
-  // check based on supported applications
-  commonAppProposalValidation(
-    params,
-    registryAppInfo,
-    [CONVENTION_FOR_ETH_ASSET_ID, contractAddresses.Token!],
-    // TODO: ^ better way to get supported token addresses?
-  );
+  const { proposal } = middlewareContext;
+  sharedProposalMiddleware(middlewareContext, contractAddresses);
   const appDef = proposal.appDefinition;
   switch (appDef) {
     case contractAddresses.DepositApp: {
@@ -133,45 +102,9 @@ const proposeMiddleware = async (
       break;
     }
     default: {
-      return;
-    }
-  }
-};
-
-const installMiddleware = async (
-  network: NetworkContext,
-  middlewareContext: InstallMiddlewareContext,
-) => {
-  const { contractAddresses } = network;
-  const { appInstance } = middlewareContext;
-  switch (appInstance.appInterface.addr) {
-    case contractAddresses.DepositApp:
-    case contractAddresses.SimpleTwoPartySwapApp:
-    case contractAddresses.SimpleSignedTransferApp:
-    case contractAddresses.SimpleLinkedTransferApp:
-    case contractAddresses.HashLockTransferApp:
-    case contractAddresses.WithdrawApp:
-    default: {
-      throw new Error("installMiddleware not implemented");
-    }
-  }
-};
-
-const takeActionMiddleware = async (
-  network: NetworkContext,
-  middlewareContext: TakeActionMiddlewareContext,
-) => {
-  const { contractAddresses } = network;
-  const { appInstance } = middlewareContext;
-  switch (appInstance.appInterface.addr) {
-    case contractAddresses.DepositApp:
-    case contractAddresses.SimpleTwoPartySwapApp:
-    case contractAddresses.SimpleSignedTransferApp:
-    case contractAddresses.SimpleLinkedTransferApp:
-    case contractAddresses.HashLockTransferApp:
-    case contractAddresses.WithdrawApp:
-    default: {
-      throw new Error("takeActionMiddleware not implemented");
+      throw new Error(
+        `Not installing app without configured validation. Cxt: ${stringify(middlewareContext)}`,
+      );
     }
   }
 };
