+++ modules/apps/src/shared/index.ts
@@ -1,2 +1,3 @@
 export * from "./registry";
 export * from "./validation";
+export * from "./middleware";
